% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{dtelegraph}
\alias{dtelegraph}
\title{The exact solution to the telegraph equation with Dirac IC}
\usage{
dtelegraph(t, x, x0, f, sp = 1)
}
\arguments{
\item{t}{time input}

\item{x}{space input}

\item{x0}{initial position of particle}

\item{f}{the initial condition of the PDE, here the Dirac IC}

\item{sp}{the speed of light}
}
\value{
list of time input, space input, and solution matrix
}
\description{
{See fundamental Green functions for telegrapher PDE.}
}
