% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo.R
\name{monteCarlo}
\alias{monteCarlo}
\title{Solve the telegrapher PDE via Kac's Monte-Carlo method}
\usage{
monteCarlo(x, f, s, sp = 1, ...)
}
\arguments{
\item{x}{grid of space coordinates to compute the solution over}

\item{f}{the initial condition to the wave-equation, as a function of \eqn{(x,...)}}

\item{s}{a collection of sample-paths of the Kac random time}

\item{sp}{speed of light}

\item{...}{other arguments to pass to the initial condition}
}
\value{
list of the time grid 't', space grid 'x' and solution matrix 'u'
}
\description{
{Monte-Carlo solver of the telegrapher PDE via averaging
Goldstein-Kac processes.}
}
