% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicitScheme.R
\name{pde}
\alias{pde}
\title{Numerical solver for a special case of the Telegrapher PDE}
\usage{
pde(x0, tt, N, M, sp = 1, xScale = 1, f = NULL)
}
\arguments{
\item{x0}{initial point in space}

\item{tt}{the time horizon}

\item{N}{the time resolution (number of time sub-intervals)}

\item{M}{the space resolution (number of space sub-intervals)}

\item{sp}{the speed of light}

\item{xScale}{scale up or down the space interval \eqn{(x_0-ct, x0+ct)}}

\item{f}{the initial condition as a function of the space coordinate}
}
\value{
list of the time-grid, space-grid, and solution matrix.
}
\description{
{A basic implicit scheme implementation for the formal
analog of the Fokker-Planck equation in Minkowski space with one space
dimension and sign convention \eqn{-, +}. The initial condition is
the dirac delta function and the initial velocity is zero.}
}
