% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePath.R
\name{plotEnsemble}
\alias{plotEnsemble}
\title{Wrapper for plotting all sample paths}
\usage{
plotEnsemble(ensemble)
}
\arguments{
\item{ensemble}{matrix of sample paths returned from \code{samplePath}}
}
\description{
{Plot multiple sample paths at once.}
}
