% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePath.R
\name{plotTelegraphSample}
\alias{plotTelegraphSample}
\title{Plot wrapper for the Telegraph-process}
\usage{
plotTelegraphSample(s, sp)
}
\arguments{
\item{s}{data.frame returned from \code{samplePath}}

\item{sp}{the speed of light}
}
\value{
null
}
\description{
{This is to plot all of the processes involved
in a single sample-path simulation: the Poisson process,
the random sign change, and the photon itself.}
}
