% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePath.R
\name{sampleEnsemble}
\alias{sampleEnsemble}
\title{Simulate a sample path of the Goldstein-Kac process}
\usage{
sampleEnsemble(tt, x0, sp = 1, ntime, npaths, startSign = NULL)
}
\arguments{
\item{tt}{length of time to simulate under}

\item{x0}{the initial position of the particle}

\item{sp}{the speed of light}

\item{ntime}{number of time sub-intervals}

\item{npaths}{number of paths to generate}

\item{startSign}{the starting direction, use NULL for random, otherwise
use \eqn{+/- 1}}
}
\value{
matrix where the first column is the time, the rest
are paths
}
\description{
{Simulate a sample path of the Goldstein-Kac
process.}
}
