% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePath.R
\name{samplePath}
\alias{samplePath}
\title{Generate a sample path of the Telegraph photon process}
\usage{
samplePath(tt, x0, sp, n, startSign = NULL)
}
\arguments{
\item{tt}{length of time to run a sample-path for}

\item{x0}{the initial position of the particle}

\item{sp}{the speed of light/velocity of the particle}

\item{n}{the number of time sub-intervals to use}

\item{startSign}{the starting direction to use +/- 1, can be NULL for random}
}
\value{
data.frame containing the time step, the Poisson process,
the random sign changing process, Kac's "random time" and finally the telegrapher process proper.
}
\description{
{A basic Euler-scheme to generate a sample path
of a Telegraph process, or also known as the Goldstein-Kac process.
It travels at a constant absolute velocity changing directions randomly,
according to a Poisson process.}
}
