% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegraphModel.R
\name{telegraph_model}
\alias{telegraph_model}
\title{Compute the PDF of the telegraph process under three methods}
\usage{
telegraph_model(
  x0,
  tt,
  sp = 1,
  ntime = 250,
  N = 400,
  M = 400,
  npaths = 50,
  xscale = 1.5
)
}
\arguments{
\item{x0}{initial point}

\item{tt}{time horizon}

\item{sp}{speed of light}

\item{ntime}{number of time sub-intervals in sample-paths}

\item{N}{number of time sub-intervals in PDE grid}

\item{M}{number of space sub-intervals in PDE grip}

\item{npaths}{number of paths in ensemble for Monte-Carlo scheme}

\item{xscale}{the scale of the region for the PDE.}
}
\value{
list of the three pdfs plus a list of the input.
}
\description{
{Compute and plot the PDFs and CDFs of the telegraph process
at a given time and initial point, using the exact semi-analytic formula, a Monte-Carlo
scheme and a numerical PDE solver.}
}
